src/main/java/jobsite/service/UserService.java

package jobsite.service;

import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
public class UserService implements UserDetailsService {

    @Override
    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
        // Implement user loading logic from your repository
        // This is just a placeholder that should be replaced with actual implementation
        throw new UsernameNotFoundException("User not found: " + username);
    }
}
